Isolate CultureInfo EH residual pin
Move the CultureInfo EH-unstructured pivot pin out of CompletenessTests so the #1135 curation PRs can merge independently.
